requirements.txt
pandas
json
datetime
pymongo
pyyaml